requirements.txt
numpy==1.14.0
imutils==0.4.6
opencv_python==3.4.1.15
paho_mqtt==1.3.1
